[#6] refactor: 댓글 관련(레포지토리) 추상화
